[Critical] Fix unserialization of dat
Previously usnerialization in OTC style was done, however the enums and basically attr evaluation is done weirdly there, so we swap and no longer have any issues with weird .dat.

The example of how weirdly it worked previously:
noMoveAnimation and Useable were interwined, and it would output an useable item as: noMoveAnimation true and useable false.